requirements.txt
easydict==1.12
gluonts==0.14.4
matplotlib==3.7.1
numpy==1.24.2
pandas==2.0.0
Pillow==9.4.0
pyrootutils==1.0.4
pytorch_lightning==2.2.0.post0
PyYAML==6.0
scikit_learn==1.3.0
scipy==1.8.1
setuptools==52.0.0
torch==2.0.1
torchmetrics==1.3.1
torchvision==0.15.2
tqdm==4.61.2
openpyxl==3.1.2
tensorboard==2.16.2
ipykernel==6.29.3
omegaconf==2.3.0
hydra-core==1.3.2
rootutils==1.0.7
rich==13.7.1
hydra-colorlog==1.2.0
docopt==0.6.2
wandb==0.16.4
timm==0.9.16
nilearn==0.10.4
deepdish==0.3.7
seaborn==0.13.2
neuroCombat==0.2.12